Refactor sendRequest to support concurrency, rate limiting, and improved error handling
Introduced worker goroutines controlled by a workerCount parameter for concurrent request processing.
Implemented rate limiting using golang.org/x/time/rate.Limiter to control bandwidth usage.
Added retry logic with exponential backoff for failed requests, up to a maximum of 2 retries.
Enhanced logging to provide detailed insights, including warnings when frequently hitting the bandwidth limit.
Removed manual rate tracking in favor of rate.Limiter for more accurate and efficient rate control.

+++ cmd/gospy/main.go
@@ -83,7 +83,7 @@ func main() {
 				Usage: "Interval between sending accumulated samples to Pyroscope",
 				Value: DefaultAccumulationPeriod,
 			},
-			&cli.IntFlag{
+			&cli.Float64Flag{
 				Name:  "rate-mb",
 				Usage: "Ingestion rate limit in MB",
 				Value: DefaultRateMB,
